only remove the required files on checkout

+++ packages/cli/src/projects/checkout.ts
@@ -10,13 +10,14 @@ import * as o from '../options';
 import * as po from './options';
 
 import type { Opts } from './options';
+import { tidyWorkflowDir } from './util';
 
 export type CheckoutOptions = Pick<
   Opts,
-  'command' | 'project' | 'workspace' | 'log'
+  'command' | 'project' | 'workspace' | 'log' | 'clean'
 >;
 
-const options = [o.log, po.workspace];
+const options = [o.log, po.workspace, po.clean];
 
 const command: yargs.CommandModule = {
   command: 'checkout <project>',
@@ -40,6 +41,8 @@ export const handler = async (options: CheckoutOptions, logger: Logger) => {
   // TODO: try to retain the endpoint for the projects
   const { project: _, ...config } = workspace.getConfig() as any;
 
+  const currentProject = workspace.getActiveProject();
+
   // get the project
   let switchProject;
   if (/\.(yaml|json)$/.test(projectIdentifier)) {
@@ -60,7 +63,11 @@ export const handler = async (options: CheckoutOptions, logger: Logger) => {
   }
 
   // delete workflow dir before expanding project
-  await rimraf(path.join(workspacePath, config.workflowRoot ?? 'workflows'));
+  if (options.clean) {
+    await rimraf(path.join(workspacePath, config.workflowRoot ?? 'workflows'));
+  } else {
+    await tidyWorkflowDir(currentProject!, switchProject);
+  }
 
   // expand project into directory
   const files: any = switchProject.serialize('fs');
